# api/api.yml
...
openapi: 3.0.4
info:
  title: Fiber Goth API
  description: API Reference for your Fiber Goth Instance
  version: 1.0.0
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        emailVerified:
          type: boolean
          default: false
          readOnly: true
        image:
          type: string
        createdAt:
          type: string
          format: date-time
          default: Generated at runtime
        updatedAt:
          type: string
          format: date-time
          default: Generated at runtime
      required:
      - name
      - email
      - createdAt
      - updatedAt
    Session:
      type: object
      properties:
        id:
          type: string
        expiresAt:
          type: string
          format: date-time
        token:
          type: string
        createdAt:
          type: string
          format: date-time
          default: Generated at runtime
        updatedAt:
          type: string
          format: date-time
        ipAddress:
          type: string
        userAgent:
          type: string
        userId:
          type: string
        activeOrganizationId:
          type: string
        activeTeamId:
          type: string
      required:
      - expiresAt
      - token
      - createdAt
      - updatedAt
      - userId
...
  securitySchemes:
    apiKeyCookie:
      type: apiKey
      in: cookie
      name: apiKeyCookie
      description: API Key authentication via cookie
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
security:
- apiKeyCookie: []
  bearerAuth: []
servers:
- url: http://localhost:3000/api/auth
...
paths:
...
  "/get-session":
    get:
      tags:
      - Default
      description: Get the current session
      operationId: getSession
      security:
      - bearerAuth: []
      parameters: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                nullable: true
                properties:
                  session:
                    "$ref": "#/components/schemas/Session"
                  user:
                    "$ref": "#/components/schemas/User"
                required:
                - session
                - user
        '400':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                - message
          description: Bad Request. Usually due to missing parameters, or invalid
            parameters.
        '401':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                required:
                - message
          description: Unauthorized. Due to missing or invalid authentication.
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
          description: Forbidden. You do not have permission to access this resource
            or to perform this action.
        '404':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
          description: Not Found. The requested resource was not found.
        '429':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
          description: Too Many Requests. You have exceeded the rate limit. Try again
            later.
        '500':
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
          description: Internal Server Error. This is a problem with the server that
            you cannot fix.
...